New files new file:   diagnostics/IRI/IOD_csv_sea.ncl new file:   diagnostics/IRI/NINO3.4_csv_sea.ncl new file:   diagnostics/IRI/RONI_csv_notify.sh new file:   diagnostics/IRI/RONI_csv_sea.ncl new file:   diagnostics/IRI/compute_std_tropical_nino.sh new file:   diagnostics/IRI/compute_std_tropical_nino_sea.ncl new file:   diagnostics/IRI/deleteme.sh new file:   diagnostics/IRI/launch_csv_IRI.sh
Updates
modified:   IC_CPS/IC_CLM/launch_forced_run_EDA.sh
modified:   IC_CPS/IC_NEMO/make_daily_anom_graph.ncl
modified:   diagnostics/C3S/ncl/ENSO_plot.ncl
modified:   diagnostics/C3S/ncl/ENSO_prob_seas_plot.ncl
modified:   postproc/APEC/launch_submit_APEC_C3S.sh
modified:   postproc/WMO/launch_push4WMO.sh
modified:   postproc/nemo/interp_ORCA2_1X1_gridT2C3S_template.sh
modified:   util/copy_SPS4extended_from_Leonardo.sh
modified:   util/copy_restart4extended_to_Leonardo.sh
renamed:    util/copy_restart4extended_to_Leonardo_single.sh -> util/copy_restart4extended_to_Leonardo_list.sh
modified:   util/launch_copy_SPS4Forecast_from_Leonardo.sh
modified:   util/launch_copy_SPS4extended_from_Leonardo.sh

+++ src/util/copy_SPS4extended_from_Leonardo.sh
@@ -47,6 +47,11 @@ then
    exit 0
 fi
 rsync -auv --rsh="sshpass -f $HOME/.sshpasswd ssh -l a07cmc00" a07cmc00@data.leonardo.cineca.it:${leo_dir_CASES}/$caso/logs/run_moredays_${caso}_DONE ${outdir}
+if [[ ! -f ${outdir}/run_moredays_${caso}_DONE ]]
+then
+# case not yet completed
+   exit 0
+fi
 rsync -auv --rsh="sshpass -f $HOME/.sshpasswd ssh -l a07cmc00" a07cmc00@data.leonardo.cineca.it:${leo_dir}/$caso $DIR_ARCHIVE
 #$cmd a07cmc00@data.leonardo.cineca.it:${leo_dir}/$caso $DIR_ARCHIVE
